Move SelectedPositions to SPI
