+++ presto-benchmark/src/main/java/io/prestosql/benchmark/HandTpchQuery6.java
@@ -24,9 +24,9 @@ import io.prestosql.operator.project.InputChannels;
 import io.prestosql.operator.project.PageFilter;
 import io.prestosql.operator.project.PageProcessor;
 import io.prestosql.operator.project.PageProjection;
-import io.prestosql.operator.project.SelectedPositions;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.block.Block;
+import io.prestosql.spi.block.SelectedPositions;
 import io.prestosql.spi.connector.ConnectorSession;
 import io.prestosql.sql.gen.PageFunctionCompiler;
 import io.prestosql.sql.planner.plan.AggregationNode.Step;
