+++ presto-main/src/main/java/io/prestosql/operator/index/TuplePageFilter.java
@@ -16,9 +16,9 @@ package io.prestosql.operator.index;
 import com.google.common.collect.ImmutableList;
 import io.prestosql.operator.project.InputChannels;
 import io.prestosql.operator.project.PageFilter;
-import io.prestosql.operator.project.SelectedPositions;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.block.Block;
+import io.prestosql.spi.block.SelectedPositions;
 import io.prestosql.spi.connector.ConnectorSession;
 import io.prestosql.spi.type.Type;
