+++ presto-main/src/main/java/io/prestosql/operator/project/GeneratedPageProjection.java
@@ -18,6 +18,7 @@ import io.prestosql.operator.Work;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.block.Block;
 import io.prestosql.spi.block.BlockBuilder;
+import io.prestosql.spi.block.SelectedPositions;
 import io.prestosql.spi.connector.ConnectorSession;
 import io.prestosql.spi.type.Type;
 import io.prestosql.sql.relational.RowExpression;
