+++ presto-main/src/main/java/io/prestosql/sql/gen/PageFunctionCompiler.java
@@ -39,11 +39,11 @@ import io.prestosql.operator.project.InputPageProjection;
 import io.prestosql.operator.project.PageFieldsToInputParametersRewriter;
 import io.prestosql.operator.project.PageFilter;
 import io.prestosql.operator.project.PageProjection;
-import io.prestosql.operator.project.SelectedPositions;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.PrestoException;
 import io.prestosql.spi.block.Block;
 import io.prestosql.spi.block.BlockBuilder;
+import io.prestosql.spi.block.SelectedPositions;
 import io.prestosql.spi.connector.ConnectorSession;
 import io.prestosql.sql.gen.LambdaBytecodeGenerator.CompiledLambda;
 import io.prestosql.sql.planner.CompilerConfig;
