+++ presto-main/src/test/java/io/prestosql/operator/project/TestPageProcessor.java
@@ -27,6 +27,7 @@ import io.prestosql.operator.Work;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.block.Block;
 import io.prestosql.spi.block.LazyBlock;
+import io.prestosql.spi.block.SelectedPositions;
 import io.prestosql.spi.block.VariableWidthBlock;
 import io.prestosql.spi.connector.ConnectorSession;
 import io.prestosql.spi.type.Type;
@@ -57,7 +58,7 @@ import static io.prestosql.operator.PageAssertions.assertPageEquals;
 import static io.prestosql.operator.project.PageProcessor.MAX_BATCH_SIZE;
 import static io.prestosql.operator.project.PageProcessor.MAX_PAGE_SIZE_IN_BYTES;
 import static io.prestosql.operator.project.PageProcessor.MIN_PAGE_SIZE_IN_BYTES;
-import static io.prestosql.operator.project.SelectedPositions.positionsRange;
+import static io.prestosql.spi.block.SelectedPositions.positionsRange;
 import static io.prestosql.spi.function.OperatorType.ADD;
 import static io.prestosql.spi.type.BigintType.BIGINT;
 import static io.prestosql.spi.type.VarcharType.VARCHAR;
